add benchmarks ; add explicit #[inline] annotations
This diff adds benchmarks to get more info regarding Issue #1.

It appears that the remaining difference between the "simple"
case and the "cargo" case is the result of a difference in
performance between using `match` and `if` for tight loops.

I suspect it's because of the way that match arms get reordered:
if I manually reorder the "if" statement, I can reproduce the
match performance.

Also added a couple #[inline] annotations in tables.rs, though
the difference in performance in my measurements is negligible.

Bumped version number to 0.1.1.