build: mark `no_std` as legacy feature flag
The `no_std` feature flag has no effect, anymore (as already documented
in `README.md`). Document this in `Cargo.toml`, but retain it for
backwards compatibility.

Note that for better cross-package composability, an `std` flag would
likely be preferred in the future, over a `no_std` flag. The former
allows multiple packages with different preferences on this feature to
be combined in a single build, while the latter does not.

Hence, suggest that `no_std` as feature flag is a legacy compatibility
flag, and will not be used in the future.