Add more canonical equivalence tests
Test that all canonically equivalent sequences
in Unicode's `NormalizationTest.txt` have the same width.
Currently no changes need to be made to the width logic
to ensure these tests pass. However, Unicode 16
is adding a few new characters that will be problematic
(the Kirat Rai vowel signs:
<https://www.unicode.org/charts/PDF/Unicode-16.0/U160-16D40.pdf>).
Adding this test in advance ensures that we won't forget
to account for these changes when the time comes.